data_pipeline/requirements.txt
# Airflow / Prefect
apache-airflow==2.7.3
prefect==2.14.11

# Data processing
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0
pyyaml==6.0.2

# API clients
requests==2.31.0
httpx==0.25.2
biopython==1.81

# Database
sqlalchemy==2.0.23
pymongo==4.6.1
psycopg2-binary==2.9.9

# Vector databases
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Data sources
python-dotenv==1.0.3
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
loguru==0.7.2
tqdm==4.66.1
python-dateutil==2.8.2

# Notebooks
jupyter==1.0.0
jupyterlab==4.0.9
notebook==7.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional GPU support
# faiss-gpu==1.7.4

